+++ generic_type_utility/CMakeLists.txt
@@ -22,10 +22,6 @@ ament_auto_add_library(${PROJECT_NAME} SHARED
   src/ros2/structure.cpp
 )
 
-ament_auto_add_executable(types
-  src/sample.cpp
-)
-
 ament_auto_add_executable(print
   src/examples/print.cpp
 )
@@ -38,8 +34,6 @@ if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   ament_lint_auto_find_test_dependencies()
   ament_add_gtest(${PROJECT_NAME}_test test/test.cpp)
-  ament_target_dependencies(${PROJECT_NAME}_test rclcpp)
-  target_include_directories(${PROJECT_NAME}_test PRIVATE include)
   target_link_libraries(${PROJECT_NAME}_test ${PROJECT_NAME})
 endif()
