+++ generic_type_utility/package.xml
@@ -11,9 +11,12 @@
   <depend>rclcpp</depend>
   <depend>yaml_cpp_vendor</depend>
 
+  <test_depend>ament_cmake_copyright</test_depend>
+  <test_depend>ament_cmake_cppcheck</test_depend>
   <test_depend>ament_cmake_gtest</test_depend>
+  <test_depend>ament_cmake_lint_cmake</test_depend>
+  <test_depend>ament_cmake_xmllint</test_depend>
   <test_depend>ament_lint_auto</test_depend>
-  <!-- <test_depend>ament_lint_common</test_depend> -->
 
   <export>
     <build_type>ament_cmake</build_type>
